Move `RandomFloatIntDataset`
